# src/backend/ContosoEnergy.Core/Entities/Job.cs
namespace ContosoEnergy.Core.Entities
{
    /// <summary>
    /// Represents a job within the ContosoEnergy domain.
    /// This entity belongs to the Domain layer and should remain
    /// free of infrastructure or persistence-specific concerns.
    /// </summary>
    public class Job(string name)
    {
        /// <summary>
        /// Primary key for the Job entity.
        /// Typically generated by the database.
        /// </summary>
        public Guid Id { get; set; } = Guid.NewGuid();

        /// <summary>
        /// The name or title of the job.
        /// Example: "Meter Reading", "Billing Run".
        /// </summary>
        public string Name { get; set; } = name ?? throw new ArgumentNullException(nameof(name));

        /// <summary>
        /// The date and time when the job was created.
        /// Useful for auditing, tracking, and historical reference.
        /// </summary>
        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
    }
}